requirements.txt
altair==6.0.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
bota==4.0.155
botasaurus==4.0.97
botasaurus-humancursor==4.0.83
botasaurus_api==4.0.10
botasaurus_driver==4.0.92
botasaurus_proxy_authentication==1.0.16
botasaurus_requests==4.0.38
brotli==1.2.0
bs4==0.0.2
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
close_chrome==4.0.40
cloudscraper==1.2.71
DateTime==6.0
Deprecated==1.3.1
docutils==0.22.4
et_xmlfile==2.0.0
gevent==26.4.0
geventhttpclient==2.3.9
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
javascript_fixes==1.1.29
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kagglehub==1.0.0
kagglesdk==0.1.18
loro==1.10.3
lxml==6.0.3
marimo==0.23.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgspec==0.21.0
narwhals==2.19.0
numpy==2.4.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.2
parso==0.8.6
pillow==12.2.0
plotly==6.7.0
protobuf==7.34.1
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.20.0
pymdown-extensions==10.21.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytweening==1.2.0
pytz==2026.1.post1
PyVirtualDisplay==3.0
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
rootutils==1.0.7
rpds-py==0.30.0
ScraperFC==4.5.0
selenium==4.43.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==1.0.0
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.5
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
wsproto==1.3.2
zope.event==6.1
zope.interface==8.3
